move bulma components into a directory
