+++ components/bulma/b-dropdown.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import bulma from '../src/bulma';
+import bulma from '../../src/bulma';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 
 export default {
