+++ components/bulma/b-notification.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import bulma from '../src/bulma'
+import bulma from '../../src/bulma'
 
 export default {
     props: {
