+++ components/navbar/loggedInUserMenu.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import BDropdown from '../b-dropdown';
+import BDropdown from '../bulma/b-dropdown';
 import Common from '~/mixins/common';
 import SettingModalLink from './settingModalLink'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
