+++ layout.vue
@@ -43,9 +43,9 @@ import { far } from '@fortawesome/free-regular-svg-icons';
 import Common from '~/mixins/common';
 import LocalDate from '~/components/localDate';
 import SearchForm from './components/searchForm';
-import BNotification from './components/b-notification';
+import BNotification from './components/bulma/b-notification';
 import JumpButtons from './components/jumpButtons';
-import BDropdown from './components/b-dropdown';
+import BDropdown from './components/bulma/b-dropdown';
 import skinLicense from './components/skinLicense';
 import bulma from './src/bulma';
 import BumaFooter from './components/footer.vue';
